Files renamed to adher to project specs
