+++ templates/express_mongo_ts/src/configs/serverConfig.ts
@@ -3,8 +3,8 @@ import type { TServerConfig } from "../types";
 
 dotenv.config();
 
-export const configs: TServerConfig = {
-  dbUrl: process.env.DB_URL || "mongodb://localhost:27017",
+export const config: TServerConfig = {
+  dbUrl: process.env.DB_URL || "mongodb://localhost:27017/mydb",
   host: process.env.HOST || "http://localhost",
   port: parseInt(process.env.PORT!) || 8000,
 }
