+++ templates/express_mongo_ts/src/types/configType.ts
