+++ templates/express_mongo_ts/src/types/userType.ts
